[oh-tab-zhwi.3.2] Decompose event-block shared helpers
